backend/requirements.txt
# FastAPI Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# HTTP & OAuth
httpx==0.25.1
requests==2.31.0
oauthlib==3.2.2

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# WebSockets (for real-time updates)
websockets==12.0

# Background Jobs (optional - for RL training)
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Existing project dependencies (already installed)
# stable-baselines3
# gymnasium
# xgboost
# mlflow
